refactor: Remove getBoardDetails re-export, update all usages to import directly
Update all files to import getBoardDetails directly from the generated
product-sizes-data file instead of re-exporting through queries.ts.
